Create SMILES analysis
This project takes a SMILES string, which encodes structural information for organic molecules, processes it, and returns some valuable information about the molecule including degree of unsaturation, molecular formula, and molecular weight. The bulk of the program however is devoted to creating an adjacency matrix that analyzes the connectivity encoded by the SMILES string. This can be complicated because an individual molecule can be written in many different ways in SMILES.